[#1481] Update preset combobox
