+++ core/resources/l10n/messages_cs.properties
@@ -1587,8 +1587,8 @@ CustomFinImport.error.badimage = Nemohu vyvodit tvar stabilizátoru z obrázku.
 CustomFinImport.description = Obrázek bude zmenen na cernobílý \n(cerná pro stabilizátor), ujistete se prosím, \u017Ee jste pou\u017Eily cernou barvu na stabilizátor \na bílou nebo svetlou barvu na pozadí. Stabilizátor \nse musí dotýkat steny obrázku, která predstavuje uchycení pro stabilizátor.
 
 
-PresetModel.lbl.select = Výber predvolby:
-PresetModel.lbl.database = Z databáze...
+PresetModel.lbl.nopreset = ?ádné predvolby
+PresetModel.lbl.selectpreset = Výber predvolby
 
 
 ! Component Preset Chooser Dialog
