+++ core/resources/l10n/messages_de.properties
@@ -1645,8 +1645,8 @@ CustomFinImport.error.badimage = Konnte keine Leitwerksform aus dem Bild erzeuge
 CustomFinImport.description = Das Bild wird intern in ein Schwarz-Weiß-Bild konvertiert (Leitwerk: schwarz). Bitte sicherstellen, dass das Leitwerk in einer dichten, dunklen Farbe ist, während der Hintergrund weiß oder sehr hell sein sollte. Das Leitwerk muss das untere Bildende berühren, da dies die Verbindungsstelle zur Rakete wird.
 
 
-PresetModel.lbl.select = Voreinstellung auswählen:
-PresetModel.lbl.database = Aus Datenbank...
+PresetModel.lbl.nopreset = Keine Voreinstellung
+PresetModel.lbl.selectpreset = Voreinstellung auswählen
 
 
 ! Component Preset Chooser Dialog
