+++ core/resources/l10n/messages_es.properties
@@ -737,8 +737,8 @@ PreferencesDialog.languages.default  = Idioma por defecto
 PreferencesDialog.lbl.language       = Idioma de la interfaz:
 PreferencesDialog.lbl.languageEffect = El idioma cambiar\u00e1 la pr\u00f3xima vez que abra OpenRocket.
 
-PresetModel.lbl.database = Desde la Base de Datos...
-PresetModel.lbl.select   = Prefabricado
+PresetModel.lbl.nopreset = No hay preselección
+PresetModel.lbl.selectpreset   = Prefabricado
 
 PrintDialog.but.previewAndPrint  = Vista previa e Imprimir
 PrintDialog.checkbox.showByStage = Mostrar por etapas
