+++ core/resources/l10n/messages_fr.properties
@@ -729,8 +729,8 @@ PreferencesDialog.languages.default  = Valeur syst\u00E8me par d\u00E9faut
 PreferencesDialog.lbl.language       = Langue du programme:
 PreferencesDialog.lbl.languageEffect = La langue sera chang\u00E9e apr\u00E8s avoir red\u00E9marr\u00E9 OpenRocket.
 
-PresetModel.lbl.database = A partir d'une base de donn\u00E9es...
-PresetModel.lbl.select   = Choisir une pi\u00E8ce pr\u00E9d\u00E9finie:
+PresetModel.lbl.nopreset = Aucune pr\u00E9s\u00E9lection
+PresetModel.lbl.selectpreset   = Choisir une pi\u00E8ce pr\u00E9d\u00E9finie:
 
 PrintDialog.but.previewAndPrint  = Pr\u00E9-visualiser et imprimer
 PrintDialog.checkbox.showByStage = Montrer par \u00E9tage
