+++ core/resources/l10n/messages_it.properties
@@ -1651,8 +1651,8 @@ CustomFinImport.error.badimage   = Non riesco a capire la forma della pinna dall
 CustomFinImport.description      = L'immagine sar\u00e0 convertita in bianco e nero internamente (nero per le pinne), cos\u00ec assicurati di usare un nero pieno per le pinne e bianco, o colore chiaro, per lo sfondo. La pinna ndeve toccare il fondo dell'immagine, che \u00e8 la base della pinna.
 
 
-PresetModel.lbl.select = Seleziona precaricati:
-PresetModel.lbl.database = Da database...
+PresetModel.lbl.nopreset = Nessuna preimpostazione
+PresetModel.lbl.selectpreset = Seleziona precaricati
 
 
 ! Component Preset Chooser Dialog
