+++ core/resources/l10n/messages_ja.properties
@@ -1713,8 +1713,8 @@ CustomFinImport.error.badimage  = \u753B\u50CF\u304B\u3089\u30D5\u30A3\u30F3\u5F
 CustomFinImport.description  = \u753B\u50CF\u306F\u5185\u90E8\u3067\u767D\u80CC\u666F\u3068\u9ED2\u7DDA\u306B\u5909\u63DB\u3055\u308C\u307E\u3059\u3002\u306A\u306E\u3067\u30D5\u30A3\u30F3\u306B\u306F\u6697\u3044\u8272\u306E\u5B9F\u7DDA\u3001\u80CC\u666F\u306B\u306F\u767D\u304B\u660E\u308B\u3044\u8272\u3092\u4F7F\u7528\u3057\u3066\u304F\u3060\u3055\u3044\u3002\u30D5\u30A3\u30F3\u306F\u753B\u50CF\u306E\u5E95\u9762\u306B\u63A5\u3057\u3066\u3044\u306A\u304F\u3066\u306F\u3044\u3051\u307E\u305B\u3093\u3001\u3053\u308C\u306F\u30D5\u30A3\u30F3\u306E\u5E95\u9762\u306B\u306A\u308A\u307E\u3059\u3002
 
 
-PresetModel.lbl.select  = Select preset
-PresetModel.lbl.database  = From database...
+PresetModel.lbl.nopreset  = \u30d7\u30ea\u30bb\u30c3\u30c8\u306a\u3057
+PresetModel.lbl.selectpreset  = \u30d7\u30ea\u30bb\u30c3\u30c8\u9078\u629e
 
 
 ! Component Preset Chooser Dialog
