+++ core/resources/l10n/messages_nl.properties
@@ -1967,8 +1967,8 @@ CustomFinImport.error.badimage = Kon de vorm van de vin niet afleiden uit het be
 CustomFinImport.description = De afbeelding wordt intern geconverteerd naar een zwart-wit afbeelding (zwart voor de vin), dus zorg ervoor dat je een effen donkere kleur gebruikt voor de vin, en wit of een lichte kleur voor de achtergrond. De vin moet de onderkant van het beeld raken, dat is de basis van de vin.
 
 
-PresetModel.lbl.select = Selecteer preset
-PresetModel.lbl.database = Uit databank...
+PresetModel.lbl.nopreset = Geen preset
+PresetModel.lbl.selectpreset = Selecteer preset
 
 DecalModel.lbl.select = <geen>
 DecalModel.lbl.choose = Van bestand...
