+++ core/resources/l10n/messages_pl.properties
@@ -1592,8 +1592,8 @@ update.dlg.latestVersion = Korzystasz z najnowszej wersji OpenRocket: %s.
  CustomFinImport.description  =  Obraz zostanie automatycznie zamieniony na czarno-bia\u0142y (statecznik w kolorze czarnym), wi\u0119c upewnij si\u0119, \u017Ce kszta\u0142t statecznika jest wype\u0142niony ciemnym kolorem, a t\u0142o jest bia\u0142e lub jasne. Podstawa statecznika musi przylega\u0107 do dolnej kraw\u0119dzi obrazu.
  
  
- PresetModel.lbl.select  =  Wybierz ustawienia:
- PresetModel.lbl.database  =  Z bazy danych...
+ PresetModel.lbl.nopreset  = Brak ustawienia
+ PresetModel.lbl.selectpreset  =  Wybierz ustawienia
  
  
  ! Component Preset Chooser Dialog
