+++ core/resources/l10n/messages_pt.properties
@@ -715,8 +715,8 @@ PreferencesDialog.languages.default  = Padr\u00e3o do sistema
 PreferencesDialog.lbl.language       = Idioma da interface:
 PreferencesDialog.lbl.languageEffect = A linguagem vai mudar na pr\u00f3xima vez que voc\u00ea iniciar o OpenRocket.
 
-PresetModel.lbl.database = \u00c0 partir do banco de dados...
-PresetModel.lbl.select   = Selecione ajustes pr\u00e9-definidos
+PresetModel.lbl.nopreset   = Sem pr\u00e9-definidos
+PresetModel.lbl.selectpreset   = Selecione ajustes pr\u00e9-definidos
 
 PrintDialog.but.previewAndPrint  = Visualiza\u00e7\u00e3o e impress\u00e3o
 PrintDialog.checkbox.showByStage = Mostrar por est\u00e1gio
