+++ core/resources/l10n/messages_uk_UA.properties
@@ -1812,8 +1812,8 @@ CustomFinImport.error.badimage = Could not deduce fin shape from image.
 CustomFinImport.description = The image will be converted internally to black and white image (black for the fin), so make sure you use a solid dark color for the fin, and white or a light color for the background. The fin must be touching the bottom of the image, which is the base of the fin.
 
 
-PresetModel.lbl.select = Select preset
-PresetModel.lbl.database = From database...
+PresetModel.lbl.nopreset = No preset
+PresetModel.lbl.selectpreset = Select preset
 
 DecalModel.lbl.select = <none>
 DecalModel.lbl.choose = From file...
