+++ core/resources/l10n/messages_zh_CN.properties
@@ -803,8 +803,8 @@ PreferencesDialog.languages.default  = \u7CFB\u7EDF\u9ED8\u8BA4
 PreferencesDialog.lbl.language       = \u754C\u9762\u8BED\u8A00: 
 PreferencesDialog.lbl.languageEffect = \u8BED\u8A00\u8BBE\u7F6E\u5C06\u5728OpenRocket\u91CD\u542F\u540E\u751F\u6548
 
-PresetModel.lbl.database = \u4ECE\u6570\u636E\u5E93...
-PresetModel.lbl.select   = \u9009\u62E9\u9884\u8BBE
+PresetModel.lbl.nopreset   = \u6ca1\u6709\u9884\u8bbe
+PresetModel.lbl.selectpreset = \u9009\u62E9\u9884\u8BBE
 
 PrintDialog.but.previewAndPrint  = \u9884\u89C8 & \u6253\u5370
 PrintDialog.checkbox.showByStage = \u6309\u7EA7\u663E\u793A
