+++ swing/src/net/sf/openrocket/gui/configdialog/RocketComponentConfig.java
@@ -22,8 +22,11 @@ import javax.swing.JSpinner;
 import javax.swing.JTabbedPane;
 import javax.swing.JTextArea;
 import javax.swing.JTextField;
+import javax.swing.SwingUtilities;
 
 import net.miginfocom.swing.MigLayout;
+import net.sf.openrocket.database.ComponentPresetDatabase;
+import net.sf.openrocket.database.DatabaseListener;
 import net.sf.openrocket.document.OpenRocketDocument;
 import net.sf.openrocket.gui.SpinnerEditor;
 import net.sf.openrocket.gui.adaptors.BooleanModel;
@@ -36,6 +39,7 @@ import net.sf.openrocket.gui.components.BasicSlider;
 import net.sf.openrocket.gui.components.StyledLabel;
 import net.sf.openrocket.gui.components.StyledLabel.Style;
 import net.sf.openrocket.gui.components.UnitSelector;
+import net.sf.openrocket.gui.dialogs.preset.ComponentPresetChooserDialog;
 import net.sf.openrocket.gui.util.GUIUtil;
 import net.sf.openrocket.gui.widgets.SelectColorButton;
 import net.sf.openrocket.l10n.Translator;
@@ -115,8 +119,19 @@ public class RocketComponentConfig extends JPanel {
 			// If the component supports a preset, show the preset selection box.
 			presetModel = new PresetModel(this, document, component);
 			presetComboBox = new JComboBox(presetModel);
+			presetComboBox.setMaximumRowCount(20);
 			presetComboBox.setEditable(false);
-			this.add(presetComboBox, "");
+			this.add(presetComboBox);
+
+			JButton selectPreset = new SelectColorButton(trans.get("PresetModel.lbl.selectpreset"));
+			selectPreset.setToolTipText(trans.get("PresetModel.lbl.selectpreset.ttip"));
+			selectPreset.addActionListener(new ActionListener() {
+				@Override
+				public void actionPerformed(ActionEvent e) {
+					selectPreset();
+				}
+			});
+			this.add(selectPreset);
 		}
 
 		tabbedPane = new JTabbedPane();
@@ -239,6 +254,24 @@ public class RocketComponentConfig extends JPanel {
 		} else {
 			multiCompEditLabel.setText("");
 		}
+	}
+
+	private void selectPreset() {
+		SwingUtilities.invokeLater(new Runnable() {
+			@Override
+			public void run() {
+				if (presetComboBox == null || presetModel == null) return;
+				((ComponentPresetDatabase) Application.getComponentPresetDao()).addDatabaseListener(presetModel);
+				ComponentPresetChooserDialog dialog =
+						new ComponentPresetChooserDialog(SwingUtilities.getWindowAncestor(RocketComponentConfig.this), component);
+				dialog.setVisible(true);
+				ComponentPreset preset = dialog.getSelectedComponentPreset();
+				if (preset != null) {
+					presetComboBox.setSelectedItem(preset);
+				}
+				((ComponentPresetDatabase) Application.getComponentPresetDao()).removeChangeListener(presetModel);
+			}
+		});
 	}
 
 	public void clearConfigListeners() {
